Make the influencers context and Filter improvement
