+++ src/components/InfluencersCards/Influencers.js
@@ -1,31 +1,11 @@
-import React, { useEffect, useState } from 'react';
+import React, { useContext } from 'react';
 import { useMediaQuery } from 'react-responsive';
 import InfluencersCards from './InfluencersCards';
 import ResponsiveInfluencersCards from './ResponsiveInfluencersCards';
-import { collection, getFirestore, onSnapshot } from 'firebase/firestore';
+import InfluencerContext from '../InfluencersContext';
 
 const Influencers = () => {
-  const [influencersData, setInfluencersData] = useState([]);
-
-  useEffect(() => {
-    const fetchData = async () => {
-      const db = getFirestore();
-      const influencersCollectionRef = collection(db, 'influencers');
-
-      const unsubscribe = onSnapshot(influencersCollectionRef, (querySnapshot) => {
-        const influencers = [];
-        querySnapshot.forEach((doc) => {
-          influencers.push({ id: doc.id, ...doc.data() });
-        });
-        setInfluencersData(influencers);
-      });
-
-      return unsubscribe;
-    };
-
-    fetchData();
-  }, []);
-
+  const influencersData = useContext(InfluencerContext);
   const isMobile = useMediaQuery({ maxWidth: 1200 });
 
   return isMobile ? (
