+++ src/index.js
@@ -3,11 +3,14 @@ import ReactDOM from 'react-dom/client';
 import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
+import { InfluencerProvider } from './components/InfluencersContext';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
   <React.StrictMode>
-    <App />
+    <InfluencerProvider>
+      <App />
+    </InfluencerProvider>
   </React.StrictMode>
 );
